20.SSDT - Modified

+++ reporting/reporting/Tables/FactMonthlyOrders.sql
@@ -3,7 +3,7 @@
 	[LocationId] INT NOT NULL, 
     [VendorId] INT NOT NULL, 
 	[Month] INT NOT NULL, 
-    [PurchaseAmount] INT NOT NULL,
+    [PurchaseAmount] MONEY NOT NULL,
     [OrderCount] INT NOT NULL, 
 	CONSTRAINT PK_FactMonthlyOrders_LocationId_VendorId_Month PRIMARY KEY CLUSTERED ([LocationId],[VendorId],[Month]),
 	CONSTRAINT [FK_FactMonthlyOrders_Location_LocationId] FOREIGN KEY ([LocationId]) REFERENCES [DimLocation]([LocationId]),
